[FEAT] BigInt for url ID allowing for more urls

+++ models/url.py
@@ -1,4 +1,4 @@
-from sqlalchemy import DateTime, Column, Integer, String, Text
+from sqlalchemy import BigInteger, Column, DateTime, String
 
 from core.db import Base
 from schemas.types import naive_utcnow
@@ -7,7 +7,7 @@ from schemas.types import naive_utcnow
 class URL(Base):
     __tablename__ = "url"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     original_url = Column(String)
     short_url = Column(String(8))
     created_at = Column(DateTime, default=naive_utcnow)
